refactor(showcase): share journal-drain barrier; document D4 proxy-capture drain step
Extract waitForJournalDrain + countCompletedTurns (+ isCompletedTurn) out of
record-d5-fixtures.mjs into a shared, importable module,
showcase/scripts/lib/journal-drain.mjs. The D5 recorder now imports the barrier
instead of defining it inline; its behavior is unchanged. The barrier unit
tests are repointed at the shared module (9/9 green).

The same capture race lives in the D4 proxy-capture flow, but D4 fixtures are
hand-authored/mirrored — there is no committable D4 capture script to wire the
barrier into (see showcase/aimock/README.md: no CLI, --record not wired up).
So D4 coverage is documentation: aimock/README.md now describes the drain step
required when proxy-capturing a D4 fixture and points at the shared helper.

The injectable fetchImpl seam is JSDoc-typed to the minimal journal-fetch
interface the barrier actually uses (not the full DOM fetch signature), so
tests inject a tiny fake without satisfying RequestInfo/Response.
